Link rules to pull request issues
Brief, oneline rules are not always obvious. Also, in the feature, some
of the rules can become out-of-date or someone new in the team can disagree
with them. It would be a good practice to always add a link to the relative
issue next to the rule.

Inspired by http://betterspecs.org/