Squash commits from feature branches before making a merge.
We should always squash commits from a feature branch before merging
into dev(elop) branch. You will gain cleaner history and it would be
easier to revert if you need to give up the change.

Do it even if the feature branch is not local, but you are collaborating
with it.

Don't do it if branch is neither a "feature branch" nor "hotfix branch".